Now we tell you your browser sucks in your native tongue.
Moved files to their appropriate folder.
